Reformat and Mavenize
